% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{fiConfInt}
\alias{fiConfInt}
\title{Title}
\usage{
fiConfInt(parameter, fidsamples, conf = 0.95)
}
\arguments{
\item{parameter}{xx}

\item{fidsamples}{xx}

\item{conf}{xx}
}
\value{
xx
}
\description{
Title
}
\examples{
xx
}
