% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filinreg.R
\name{filinreg}
\alias{filinreg}
\alias{filinregR}
\title{Fiducial sampler for linear regression model}
\usage{
filinregR(
  formula,
  data = NULL,
  distr = "student",
  df = Inf,
  L = 10L,
  lucky = FALSE
)

filinreg(
  formula,
  data = NULL,
  distr = "student",
  df = Inf,
  L = 10L,
  lucky = TRUE
)
}
\arguments{
\item{formula}{xx}

\item{data}{xx}

\item{distr}{the distribution of the error terms, \code{"student"} or
\code{"logistic"}}

\item{df}{degrees of freedom of the Student distribution if
\code{distr = "student"}}

\item{L}{xx}

\item{lucky}{xx}
}
\value{
xxx
}
\description{
ddd
}
\examples{
xxx

}
