% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.R
\name{filinregPredictive}
\alias{filinregPredictive}
\title{Title}
\usage{
filinregPredictive(fidsamples, newdata)
}
\arguments{
\item{fidsamples}{xx}

\item{newdata}{xx}
}
\value{
xx
}
\description{
Title
}
\examples{
xx
}
